cleaning

+++ main.py
@@ -43,7 +43,7 @@ def main_function(meal, budget):
 
 
 if __name__ == "__main__":
-    meal = int(input(ru.QUESTION_1))  # local in ()
-    portions = int(input(ru.QUESTION_2))  # local in ()
-    budget = int(input(ru.QUESTION_3))  # local in ()
+    meal = float(input(ru.QUESTION_1))  # local in ()
+    portions = float(input(ru.QUESTION_2))  # local in ()
+    budget = float(input(ru.QUESTION_3))  # local in ()
     main_function(meal, budget)
